git-svn-id: https://scm.nearinfinity.com/svn/incubator/super-lucene/trunk@16014 5c733ba6-9949-11dd-800e-6d5850b5fe75

+++ src/com/nearinfinity/blur/BlurQuery.java
@@ -16,6 +16,6 @@ public interface BlurQuery {
 	long searchFast(String query, String filter);
 	long searchFast(String query, String filter, long minimum);
 	
-	List<BlurHit> searchFast(String query, String filter, int starting, int fetch);
+	List<BlurHit> searchFast(String query, String filter, long starting, int fetch);
 
 }
